+++ DTCD-StyleSystem/src/stories/BaseExpander.stories.js
@@ -17,6 +17,16 @@ export default {
       type: 'string',
       description: 'Slot "icon"',
     },
+    theme: { 
+      control: {
+        type: 'multi-select',
+      },
+      options: [
+        '',
+        'with_border-bottom',
+      ],
+      description: 'Configuration view component',
+    },
     open: {
       type: 'boolean',
       description: 'Expander is open or not',
@@ -35,6 +45,7 @@ window.customElements.define(NAME_COMPONENT, BaseExpander);
 // More on component templates: https://storybook.js.org/docs/web-components/writing-stories/introduction#using-args
 const Template = (args) => {
   const {
+    theme = [],
     open,
     summarySlot,
     iconSlot,
@@ -46,11 +57,8 @@ const Template = (args) => {
   expander.innerHTML += iconSlot;
   expander.innerHTML += defaultSlot;
 
-  if (open) {
-    expander.setAttribute('open', 'open');
-  } else {
-    expander.removeAttribute('open');
-  }
+  expander.open = open ? true : false;
+  expander.theme = theme?.length ? theme : [];
 
   return expander;
 };
@@ -58,8 +66,18 @@ const Template = (args) => {
 export const DefaultExpander = Template.bind({});
 // More on args: https://storybook.js.org/docs/web-components/writing-stories/args
 DefaultExpander.args = {
-  summarySlot: '<div slot="summary">Summary slot</div>',
+  summarySlot: '<div slot="summary">Default expander</div>',
   iconSlot: '',
   defaultSlot: 'Default slot. Lorem ipsum dolor sit amet consectetur adipisicing elit. Veritatis veniam, dolor explicabo dolore nobis, minima rerum obcaecati eius fuga fugit, nemo consequuntur nesciunt itaque necessitatibus repellendus recusandae porro soluta consequatur?',
+  theme: [],
+  open: false,
+};
+
+export const ExpanderWithBottomBorder = Template.bind({});
+ExpanderWithBottomBorder.args = {
+  summarySlot: '<div slot="summary">Expander with bottom border</div>',
+  iconSlot: '',
+  defaultSlot: 'Default slot. Lorem ipsum dolor sit amet consectetur adipisicing elit. Veritatis veniam, dolor explicabo dolore nobis, minima rerum obcaecati eius fuga fugit, nemo consequuntur nesciunt itaque necessitatibus repellendus recusandae porro soluta consequatur?',
+  theme: ['with_border-bottom'],
   open: false,
 };